Making jquery optional
jQuery is now an optional library for Backbone.
This update follows the convention used in backbone.js which
places `require('jquery')` inside a try/catch block.

A similar technique to that described here can be used
to avoid loading jquery:
https://github.com/jashkenas/backbone/wiki/Using-Backbone-without-jQuery#without-jquery

The only usage of jQuery within this package was to create
a promise in backbone.iosync.js. If you do not include jQuery
in your project, it will instead look for an ES2015 Promise
function on the global scope.

Also added an easier way to use this package via npm:
```
var Backbone = require('backbone'),
    iobind = require('backbone.iobind');

iobind.bind();
iobind.sync();
```